cleaned up reservation components
