+++ front-end/src/Reservations/ReservationList.js
@@ -1,7 +1,10 @@
-import { useHistory } from "react-router";
+import { useHistory } from "react-router-dom";
 import { today, previous, next } from "../utils/date-time";
+import ReservationCards from "./ReservationCards";
+
 function ReservationList({ date, reservations }) {
   const history = useHistory();
+
   const handleTodaysDate = () => {
     history.push(`dashboard?date=${today()}`);
   };
@@ -38,46 +41,19 @@ function ReservationList({ date, reservations }) {
       </div>
     );
   };
-  const reservationCards = reservations
-    .filter(({ reservation_date }) => reservation_date === date)
-    .map(
-      ({
-        reservation_id,
-        first_name,
-        last_name,
-        mobile_number,
-        reservation_date,
-        reservation_time,
-        people,
-      }) => {
-        return (
-          <div
-            className="card border-success mb-3"
-            style={{ maxWidth: "40rem" }}
-            key={reservation_id}
-          >
-            <div className="card-header bg-transparent border-success">
-              {reservation_date}
-            </div>
-            <div className="card-body text-success">
-              <h5 className="card-title">{reservation_time}</h5>
-              <p className="card-text">
-                {first_name} {last_name}
-              </p>
-              <p className="card-text">Number of people: {people}</p>
-            </div>
-            <div className="card-footer bg-transparent border-success">
-              {mobile_number}
-            </div>
-          </div>
-        );
-      }
-    );
 
+  const active = reservations.map((reservation) => {
+    return (
+      <ReservationCards
+        key={reservation.reservation_id}
+        reservation={reservation}
+      />
+    );
+  });
   return (
     <div>
-      <div className="mb-3"> {controlButtons()}</div>
-      <div className="card-columns">{reservationCards}</div>
+      <div className="mb-3">{controlButtons()}</div>
+      <div>{active}</div>
     </div>
   );
 }
