+++ front-end/src/dashboard/Dashboard.js
@@ -3,6 +3,7 @@ import { listReservations } from "../utils/api";
 import ErrorAlert from "../layout/ErrorAlert";
 import useQuery from "../utils/useQuery";
 import ReservationList from "../Reservations/ReservationList";
+
 /**
  * Defines the dashboard page.
  * @param date
@@ -32,7 +33,7 @@ function Dashboard({ date }) {
       <main>
         <h1>Dashboard</h1>
         <div className="d-md-flex mb-3">
-          <h4 className="mb-0">Reservations for {date}</h4>
+          <h4 className="mb-0">Reservations for: {date}</h4>
         </div>
         <ReservationList date={date} reservations={reservations} />
         <ErrorAlert error={reservationsError} />
