+++ front-end/src/layout/Routes.js
@@ -1,10 +1,10 @@
 import React from "react";
 
 import { Redirect, Route, Switch } from "react-router-dom";
-import Reservations from "../Reservations";
 import Dashboard from "../dashboard/Dashboard";
 import NotFound from "./NotFound";
 import { today } from "../utils/date-time";
+import CreateReservation from "../Reservations/CreateReservation";
 
 /**
  * Defines all the routes for the application.
@@ -26,7 +26,7 @@ function Routes() {
         <Dashboard date={today()} />
       </Route>
       <Route path="/reservations/new">
-        <Reservations />
+        <CreateReservation />
       </Route>
       <Route>
         <NotFound />
